requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
openai==1.1.0
langchain==0.0.300
transformers==4.34.0
sentence-transformers==2.2.2
scikit-learn==1.2.2
nltk==3.8.1
uvicorn==0.22.0
requests==2.31.0
